Use Kotlin Endpoint in HttpKit. now need to fix the use places in other frameworks like BrowserNetworking
