+++ CoreHttpKit/src/commonMain/kotlin/ParametersEncodingDestination.kt
@@ -4,14 +4,14 @@ package org.cottonweb.CoreHttpKit
  * URLSearchParams analog from Apple and the value could be optional
  * On the other hand, Parameters can store only non optional pairs
  * */
-data class URLQueryItem(val name: String, val value: String)
+data class URLQueryPair(val name: String, val value: String)
 
 sealed class ParametersEncodingDestination {
-    class QueryString(val items: Array<URLQueryItem>) : ParametersEncodingDestination()
+    class QueryString(val items: Array<URLQueryPair>) : ParametersEncodingDestination()
     class HttpBodyJSON(val keyToValue: Map<String, Any>) : ParametersEncodingDestination()
     class HttpBody(val encodedData: ByteArray) : ParametersEncodingDestination()
 
-    val contentTypeHttpHeader: HTTPHeader
+    internal val contentTypeHttpHeader: HTTPHeader
         get() {
             val contentType: ContentTypeValue = when (this) {
                 is QueryString -> ContentTypeValue.Url
