+++ catowser/HttpKit/ClosureVoidWrappers.swift
@@ -9,16 +9,17 @@
 #if canImport(Combine)
 import Combine
 #endif
+import CoreHttpKit
 
 extension HttpKit {
     // gryphon ignore
     public class ClosureVoidWrapper<Server: ServerDescription>: Hashable {
         public var closure: (Result<Void, HttpKit.HttpError>) -> Void
         /// Don't need to use endpoint here, but it is needed to create unique hash value for the closure
-        let endpoint: VoidEndpoint<Server>
+        let endpoint: Endpoint<Server>
         
         public init(_ closure: @escaping (Result<Void, HttpKit.HttpError>) -> Void,
-                    _ endpoint: VoidEndpoint<Server>) {
+                    _ endpoint: Endpoint<Server>) {
             self.closure = closure
             self.endpoint = endpoint
         }
@@ -43,10 +44,10 @@ extension HttpKit {
     public class CombinePromiseVoidWrapper<Server: ServerDescription>: Hashable {
         public var promise: Future<Void, HttpKit.HttpError>.Promise
         /// Don't need to use endpoint here, but it is needed to create unique hash value for the closure
-        let endpoint: VoidEndpoint<Server>
+        let endpoint: Endpoint<Server>
         
         public init(_ promise: @escaping Future<Void, HttpKit.HttpError>.Promise,
-                    _ endpoint: VoidEndpoint<Server>) {
+                    _ endpoint: Endpoint<Server>) {
             self.promise = promise
             self.endpoint = endpoint
         }
