+++ catowser/HttpKit/Extensions/HttpKotlinTypes+Extensions.swift
@@ -1,0 +1,50 @@
+//
+//  HttpKotlinTypes+Extensions.swift
+//  HttpKit
+//
+//  Created by Andrei Ermoshin on 4/17/22.
+//  Copyright © 2022 andreiermoshin. All rights reserved.
+//
+
+import CoreHttpKit
+import Combine
+
+extension Set where Element == CoreHttpKit.HTTPHeader {
+    var dictionary: [String: String] {
+        var dictionary = [String: String]()
+        for header in self {
+            dictionary[header.key] = header.value
+        }
+        return dictionary
+    }
+}
+
+extension HTTPRequestInfo {
+    var urlRequest: URLRequest? {
+        guard let url = URL(string: rawURL) else {
+            return nil
+        }
+        
+        let portedTimeoutValue = TimeInterval(requestTimeout)
+        
+        var request = URLRequest(url: url,
+                                 cachePolicy: .reloadIgnoringLocalAndRemoteCacheData,
+                                 timeoutInterval: portedTimeoutValue)
+        request.httpMethod = method.stringValue
+        request.allHTTPHeaderFields = headers.dictionary
+        
+        guard let encodedData = httpBody else {
+            return request
+        }
+        request.httpBody = ByteArrayNativeUtils.companion.convert(byteArray: encodedData)
+        return request
+    }
+}
+
+extension KotlinArray where T == URLQueryPair {
+    static var empty: KotlinArray {
+        return .init(size: 0) { _ in
+            return nil
+        }
+    }
+}
