+++ catowser/HttpKit/HttpClient+Combine.swift
@@ -10,12 +10,13 @@ import Foundation
 #if canImport(Combine)
 import Combine
 #endif
+import CoreHttpKit
 
 extension HttpKit.Client {
     public typealias ResponseFuture<T> = Deferred<Publishers.HandleEvents<Future<T, HttpKit.HttpError>>>
     
     // gryphon ignore
-    public func cMakeRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+    public func cMakeRequest<T, B: HTTPAdapter>(for endpoint: Endpoint<Server>,
                                                 withAccessToken accessToken: String?,
                                                 transport adapter: B,
                                                 subscriber: HttpKit.Subscriber<T, Server>) -> ResponseFuture<T>
