+++ catowser/HttpKit/HttpClient+Kotlin.swift
@@ -9,7 +9,7 @@
 import CoreHttpKit
 
 extension HttpKit.Client {
-    public func kMakeRequest<T, B: HTTPAdapter>(for endpoint: Endpoint<T, Server>,
+    public func makeRequest<T, B: HTTPAdapter>(for endpoint: Endpoint<Server>,
                                                 withAccessToken accessToken: String?,
                                                 transport adapter: B) where B.Response == T, B.Server == Server {
         let requestInfo = endpoint.request(server: server,
@@ -23,36 +23,38 @@ extension HttpKit.Client {
         let codes = T.successCodes
         adapter.performRequest(httpRequest, sucessCodes: codes)
     }
-}
-
-private extension Set where Element == CoreHttpKit.HTTPHeader {
-    var dictionary: [String: String] {
-        var dictionary = [String: String]()
-        for header in self {
-            dictionary[header.key] = header.value
+    
+    // MARK: - Clear RX capable functions without dependencies
+    
+    public func makeRxRequest<T, B: HTTPRxAdapter>(for endpoint: Endpoint<Server>,
+                                                   withAccessToken accessToken: String?,
+                                                   transport adapter: B) where B.Response == T, B.Server == Server {
+        let requestInfo = endpoint.request(server: server,
+                                           requestTimeout: Int64(httpTimeout),
+                                           accessToken: accessToken)
+        guard let httpRequest = requestInfo.urlRequest else {
+            let result: HttpTypedResult<T> = .failure(.failedKotlinRequestConstruct)
+            adapter.wrapperHandler()(result)
+            return
         }
-        return dictionary
+        
+        let codes = T.successCodes
+        adapter.performRequest(httpRequest, sucessCodes: codes)
     }
-}
-
-private extension HTTPRequestInfo {
-    var urlRequest: URLRequest? {
-        guard let url = URL(string: rawURL) else {
-            return nil
+    
+    public func makeRxVoidRequest<B: HTTPRxVoidAdapter>(for endpoint: Endpoint<Server>,
+                                                        withAccessToken accessToken: String?,
+                                                        transport adapter: B) where B.Server == Server {
+        let requestInfo = endpoint.request(server: server,
+                                           requestTimeout: Int64(httpTimeout),
+                                           accessToken: accessToken)
+        guard let httpRequest = requestInfo.urlRequest else {
+            let result: Result<Void, HttpKit.HttpError> = .failure(.failedKotlinRequestConstruct)
+            adapter.wrapperHandler()(result)
+            return
         }
         
-        let portedTimeoutValue = TimeInterval(requestTimeout)
-        
-        var request = URLRequest(url: url,
-                                 cachePolicy: .reloadIgnoringLocalAndRemoteCacheData,
-                                 timeoutInterval: portedTimeoutValue)
-        request.httpMethod = method.stringValue
-        request.allHTTPHeaderFields = headers.dictionary
-        
-        guard let encodedData = httpBody else {
-            return request
-        }
-        request.httpBody = ByteArrayNativeUtils.companion.convert(byteArray: encodedData)
-        return request
+        let codes = VoidResponse.successCodes
+        adapter.performVoidRequest(httpRequest, sucessCodes: codes)
     }
 }
