+++ catowser/HttpKit/HttpClient.swift
@@ -70,95 +70,5 @@ extension HttpKit {
                 return
             }
         }
-        
-        // MARK: - Clear RX capable functions without dependencies
-        
-        public func makeRxRequest<T, B: HTTPRxAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                       withAccessToken accessToken: String?,
-                                                       transport adapter: B) where B.Response == T, B.Server == Server {
-            
-            guard let url = endpoint.url(relatedTo: self.server) else {
-                let result: HttpTypedResult<T> = .failure(.failedConstructUrl)
-                adapter.wrapperHandler()(result)
-                return
-            }
-            let httpRequest: URLRequest
-            do {
-                httpRequest = try endpoint.request(url,
-                                                   httpTimeout: self.httpTimeout,
-                                                   jsonEncoder: jsonEncoder,
-                                                   accessToken: accessToken)
-            } catch let error as HttpKit.HttpError {
-                let result: HttpTypedResult<T> = .failure(error)
-                adapter.wrapperHandler()(result)
-                return
-            } catch {
-                let result: HttpTypedResult<T> = .failure(.httpFailure(error: error))
-                adapter.wrapperHandler()(result)
-                return
-            }
-            
-            let codes = T.successCodes
-            adapter.performRequest(httpRequest, sucessCodes: codes)
-        }
-        
-        public func makeRxVoidRequest<B: HTTPRxVoidAdapter>(for endpoint: HttpKit.VoidEndpoint<Server>,
-                                                            withAccessToken accessToken: String?,
-                                                            transport adapter: B) where B.Server == Server {
-            guard let url = endpoint.url(relatedTo: self.server) else {
-                let result: Result<Void, HttpKit.HttpError> = .failure(.failedConstructUrl)
-                adapter.wrapperHandler()(result)
-                return
-            }
-            
-            let httpRequest: URLRequest
-            do {
-                httpRequest = try endpoint.request(url,
-                                                   httpTimeout: self.httpTimeout,
-                                                   jsonEncoder: jsonEncoder,
-                                                   accessToken: accessToken)
-            } catch let error as HttpKit.HttpError {
-                let result: Result<Void, HttpKit.HttpError> = .failure(error)
-                adapter.wrapperHandler()(result)
-                return
-            } catch {
-                let result: Result<Void, HttpKit.HttpError> = .failure(.httpFailure(error: error))
-                adapter.wrapperHandler()(result)
-                return
-            }
-            
-            let codes = HttpKit.VoidResponse.successCodes
-            adapter.performVoidRequest(httpRequest, sucessCodes: codes)
-        }
-        
-        // MARK: - Clear functions without dependencies
-        
-        public func makeRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                   withAccessToken accessToken: String?,
-                                                   transport adapter: B) where B.Response == T, B.Server == Server {
-            guard let url = endpoint.url(relatedTo: self.server) else {
-                let result: HttpTypedResult<T> = .failure(.failedConstructUrl)
-                adapter.wrapperHandler()(result)
-                return
-            }
-            let httpRequest: URLRequest
-            do {
-                httpRequest = try endpoint.request(url,
-                                                   httpTimeout: self.httpTimeout,
-                                                   jsonEncoder: jsonEncoder,
-                                                   accessToken: accessToken)
-            } catch let error as HttpKit.HttpError {
-                let result: HttpTypedResult<T> = .failure(error)
-                adapter.wrapperHandler()(result)
-                return
-            } catch {
-                let result: HttpTypedResult<T> = .failure(.httpFailure(error: error))
-                adapter.wrapperHandler()(result)
-                return
-            }
-            
-            let codes = T.successCodes
-            adapter.performRequest(httpRequest, sucessCodes: codes)
-        }
     }
 }
