--- catowser/HttpKit/HttpKitHeaders.swift
@@ -1,90 +1,0 @@
-//
-//  HttpKitHeaders.swift
-//  HttpKit
-//
-//  Created by Andrei Ermoshin on 10/12/19.
-//  Copyright © 2019 andreiermoshin. All rights reserved.
-//
-
-// gryphon output: ../CoreHttpKit/src/nativeMain/kotlin/HttpKitHeaders.kt
-// gryphon insert: package org.cottonweb.CoreHttpKit
-
-import Foundation
-
-/// TODO: think about namespace, probably we need to remove it to be able to convert this to Kotlin
-
-// gryphon ignore
-extension HttpKit {
-    public enum HttpHeader: Equatable, Hashable {
-        case contentType(ContentType)
-        case contentLength(Int)
-        case accept(ContentType)
-        case authorization(token: String)
-        
-        public var key: String {
-            switch self {
-            case .contentType:
-                return "Content-Type"
-            case .contentLength:
-                return "Content-Length"
-            case .accept:
-                return "Accept"
-            case .authorization:
-                return "Authorization"
-            }
-        }
-        
-        public var value: String {
-            switch self {
-            case .contentType(let type):
-                return type.rawValue
-            case .contentLength(let length):
-                return "\(length)"
-            case .accept(let type):
-                return type.rawValue
-            case .authorization(token: let token):
-                // setup different ways for authorization, not only Bearer
-                return "Bearer \(token)"
-            }
-        }
-        
-        public static func == (lhs: HttpKit.HttpHeader, rhs: HttpKit.HttpHeader) -> Bool {
-            switch (lhs, rhs) {
-            case (.contentType(let lType), contentType(let rType)):
-                return lType == rType
-            case (.contentLength(let lLength), contentLength(let rLength)):
-                return lLength == rLength
-            case (.accept(let lType), accept(let rType)):
-                return lType == rType
-            case (.authorization(token: let lToken), .authorization(token: let rToken)):
-                return lToken == rToken
-            default:
-                return false
-            }
-        }
-        
-        public func hash(into hasher: inout Hasher) {
-            hasher.combine(key)
-            hasher.combine(value)
-        }
-    }
-    
-    public enum ContentType: String {
-        case json = "application/json"
-        /// The following type is used to indicate that the response will contain search suggestions.
-        /// Link: [doc](http://www.opensearch.org/Specifications/OpenSearch/Extensions/Suggestions/1.0)
-        case jsonSuggestions = "application/x-suggestions+json"
-        case url = "application/x-www-form-urlencoded"
-        case html = "text/html"
-    }
-}
-
-extension Array where Element == HttpKit.HttpHeader {
-    var dictionary: [String: String] {
-        var dictionary = [String: String]()
-        for header in self {
-            dictionary[header.key] = header.value
-        }
-        return dictionary
-    }
-}
